requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.25.2
google-generativeai==0.3.2
